RemoteHealthBackend-main/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
python-socketio==5.10.0
python-dotenv==1.0.0
pydantic==2.4.2
pydantic-settings==2.0.3
email-validator==2.1.0.post1
psycopg2-binary==2.9.10
requests==2.31.0 